--- src/main/java/roomescape/login/service/LoginService.java
@@ -1,43 +1,0 @@
-package roomescape.login.service;
-
-import io.jsonwebtoken.Jwts;
-import io.jsonwebtoken.security.Keys;
-import java.util.Objects;
-import javax.naming.AuthenticationException;
-import org.springframework.stereotype.Service;
-import roomescape.exceptions.NotFoundException;
-import roomescape.login.dto.LoginRequest;
-import roomescape.member.domain.Email;
-import roomescape.member.domain.Member;
-import roomescape.member.domain.Password;
-import roomescape.member.repository.MemberRepository;
-
-@Service
-public class LoginService {
-
-    private static final String SECRET_KEY = "Yn2kjibddFAWtnPJ2AFlL8WXmohJMCvigQggaEypa5E";
-    private final MemberRepository memberRepository;
-
-    public LoginService(MemberRepository memberRepository) {
-        this.memberRepository = memberRepository;
-    }
-
-    public String createLoginToken(LoginRequest loginRequest) throws AuthenticationException {
-        Member member = memberRepository.findByEmail(new Email(loginRequest.email()))
-                .orElseThrow(() -> new NotFoundException("존재하지 않는 회원입니다."));
-
-        if (Objects.equals(member.getPassword(), new Password(loginRequest.password()))) {
-            return parseToToken(member);
-        }
-        throw new AuthenticationException("비밀번호가 일치하지 않습니다.");
-    }
-
-    private String parseToToken(Member member) {
-        return Jwts.builder()
-                .setSubject(member.getId().toString())
-                .claim("name", member.getName().name())
-                .claim("email", member.getEmail().email())
-                .signWith(Keys.hmacShaKeyFor(SECRET_KEY.getBytes()))
-                .compact();
-    }
-}
