+++ src/main/java/roomescape/member/service/MemberService.java
@@ -4,11 +4,15 @@ import io.jsonwebtoken.JwtException;
 import io.jsonwebtoken.Jwts;
 import io.jsonwebtoken.security.Keys;
 import java.util.List;
+import java.util.Objects;
 import javax.naming.AuthenticationException;
 import org.springframework.stereotype.Service;
 import roomescape.exceptions.NotFoundException;
+import roomescape.login.dto.LoginRequest;
+import roomescape.member.domain.Email;
 import roomescape.member.domain.LoginMember;
 import roomescape.member.domain.Member;
+import roomescape.member.domain.Password;
 import roomescape.member.domain.Role;
 import roomescape.member.dto.LoginMemberRequest;
 import roomescape.member.dto.MemberNameResponse;
@@ -37,6 +41,25 @@ public class MemberService {
                 .stream()
                 .map(MemberResponse::new)
                 .toList();
+    }
+
+    public String createMemberToken(LoginRequest loginRequest) throws AuthenticationException {
+        Member member = memberRepository.findByEmail(new Email(loginRequest.email()))
+                .orElseThrow(() -> new NotFoundException("존재하지 않는 회원입니다."));
+
+        if (Objects.equals(member.getPassword(), new Password(loginRequest.password()))) {
+            return parseToToken(member);
+        }
+        throw new AuthenticationException("비밀번호가 일치하지 않습니다.");
+    }
+
+    private String parseToToken(Member member) {
+        return Jwts.builder()
+                .setSubject(member.getId().toString())
+                .claim("name", member.getName().name())
+                .claim("email", member.getEmail().email())
+                .signWith(Keys.hmacShaKeyFor(SECRET_KEY.getBytes()))
+                .compact();
     }
 
     public MemberNameResponse getMemberNameResponseByToken(String token) throws AuthenticationException {
